zhrtvc/requirements.txt
cycler
sounddevice
tensorflow<=1.15.2
matplotlib
umap_learn
phkit>=0.2.2
tqdm
numba==0.48
aukit>=1.4.3
librosa
pydub
visdom
webrtcvad_wheels
numpy
scipy
multiprocess
SIP
PyQt5<=5.15.0
PyYAML
scikit_learn
torch
tensorboardX
matplotlib<=3.1.1
setproctitle
